requirements.txt
librosa~=0.7.2
numpy~=1.17.2
matplotlib~=3.1.1
scipy~=1.3.1
gunicorn~=20.0.4
django-heroku~=0.3.1
django-environ~=0.4.5
resampy~=0.2.2
numba~=0.45.1
pydub~=0.23.1